SWQL-78 | Fix binary and varbinary data type
Changed binary and varbinary datatypes to have a type of GraphQL List of Int. This makes them work out of the box with the express
graphql server, but is still a little kludgy as to create a value you have to wrap it in an array and when you get the column value back it's wrapped in an array
Also added unit tests for changes.